+++ app/src/main/java/ksnd/open/hiraganaconverter/view/dialog/SelectLanguageDialog.kt
@@ -23,8 +23,8 @@ import androidx.compose.ui.window.Dialog
 import androidx.compose.ui.window.DialogProperties
 import androidx.hilt.navigation.compose.hiltViewModel
 import ksnd.open.hiraganaconverter.R
-import ksnd.open.hiraganaconverter.view.parts.card.LanguageCard
 import ksnd.open.hiraganaconverter.view.parts.button.BottomCloseButton
+import ksnd.open.hiraganaconverter.view.parts.card.LanguageCard
 import ksnd.open.hiraganaconverter.viewmodel.PreviewSelectLanguageViewModel
 import ksnd.open.hiraganaconverter.viewmodel.SelectLanguageViewModel
 import ksnd.open.hiraganaconverter.viewmodel.SelectLanguageViewModelImpl
