+++ app/src/main/java/ksnd/open/hiraganaconverter/view/parts/card/CovertHistoryCard.kt
@@ -1,0 +1,107 @@
+package ksnd.open.hiraganaconverter.view.parts.card
+
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.wrapContentHeight
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.outlined.Close
+import androidx.compose.material3.FilledTonalIconButton
+import androidx.compose.material3.IconButtonDefaults
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedCard
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.scale
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.ColorFilter
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import ksnd.open.hiraganaconverter.view.rememberButtonScaleState
+
+@Composable
+fun ConvertHistoryCard(
+    beforeText: String,
+    time: String,
+    onClick: () -> Unit,
+    onDeleteClick: () -> Unit,
+) {
+    val buttonScaleState = rememberButtonScaleState()
+    OutlinedCard(
+        modifier = Modifier
+            .padding(top = 4.dp, start = 8.dp, end = 8.dp)
+            .wrapContentHeight()
+            .fillMaxWidth()
+            .scale(scale = buttonScaleState.animationScale.value)
+            .clickable(
+                interactionSource = buttonScaleState.interactionSource,
+                indication = null,
+                onClick = onClick,
+            ),
+        border = BorderStroke(width = 2.dp, color = MaterialTheme.colorScheme.secondary),
+    ) {
+        Row(verticalAlignment = Alignment.Top) {
+            Column(
+                modifier = Modifier
+                    .padding(all = 8.dp)
+                    .weight(1f),
+            ) {
+                ConvertHistoryCardTimeText(timeText = time)
+                Text(
+                    text = beforeText,
+                    maxLines = 2,
+                    minLines = 2,
+                    style = MaterialTheme.typography.bodyMedium,
+                    color = MaterialTheme.colorScheme.primary,
+                    overflow = TextOverflow.Ellipsis,
+                )
+            }
+            FilledTonalIconButton(
+                onClick = onDeleteClick,
+                colors = IconButtonDefaults.iconButtonColors(
+                    containerColor = Color.Transparent,
+                ),
+            ) {
+                Image(
+                    imageVector = Icons.Outlined.Close,
+                    contentDescription = "delete convert history",
+                    colorFilter = ColorFilter.tint(MaterialTheme.colorScheme.onSurface),
+                )
+            }
+        }
+    }
+}
+
+@Composable
+private fun ConvertHistoryCardTimeText(timeText: String) {
+    Row(
+        verticalAlignment = Alignment.CenterVertically,
+    ) {
+        Text(
+            modifier = Modifier.padding(end = 8.dp),
+            text = timeText,
+            style = MaterialTheme.typography.bodySmall,
+            color = MaterialTheme.colorScheme.secondary,
+        )
+        Spacer(modifier = Modifier.weight(1f))
+    }
+}
+
+@Preview
+@Composable
+private fun PreviewConvertHistoryCard() {
+    ConvertHistoryCard(
+        beforeText = "変換前文章",
+        time = "2022/12/29 18:19",
+        onClick = {},
+        onDeleteClick = {},
+    )
+}
